distributed system/requirements-simulator.txt
aiohttp==3.9.1
asyncio==3.4.3
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-instrumentation-logging==0.42b0
opentelemetry-semantic-conventions==0.42b0 